refactor: modularize equipmentListUtils into focused utility files
Split 960-line equipmentListUtils.ts into 7 focused files:
- equipmentConstants.ts: shared equipment ID constants
- jumpJetUtils.ts: jump jet equipment helpers
- internalStructureSlotUtils.ts: internal structure slot helpers
- armorSlotUtils.ts: armor slot equipment helpers
- heatSinkEquipmentUtils.ts: heat sink equipment helpers
- enhancementEquipmentUtils.ts: MASC/TSM/Supercharger helpers
- targetingComputerUtils.ts: targeting computer helpers

Main file (equipmentListUtils.ts) is now a 65-line aggregator that
re-exports all functions, maintaining backward compatibility.
